Extract structures for uniform variables into a separate source file.
